+++ AccountingScholarships.Domain/Entities/Real/university/Edu_StudentCourses.cs
@@ -13,9 +13,9 @@ namespace AccountingScholarships.Domain.Entities.Real.university
         public int SemesterCourseID { get; set; } // FK
         public string RegisteredBy { get; set; }
         public DateTime RegisteredOn { get; set; }
-        public double? Grade1 { get; set; }
-        public double? Grade2 { get; set; }
-        public double? ExamGrade { get; set; }
+        public float? Grade1 { get; set; }
+        public float? Grade2 { get; set; }
+        public float? ExamGrade { get; set; }
         public string? LetterGrade { get; set; }
         public string LastUpdatedBy { get; set; }
         public DateTime LastUpdatedOn { get; set; }
@@ -25,8 +25,8 @@ namespace AccountingScholarships.Domain.Entities.Real.university
         public int? prevID { get; set; }
         public int? MissingPercentage { get; set; }
         public bool MissingFailure { get; set; }
-        public double? UnsubmittedGrade1 { get; set; }
-        public double? UnsubmittedGrade2 { get; set; }
+        public float? UnsubmittedGrade1 { get; set; }
+        public float? UnsubmittedGrade2 { get; set; }
         public bool Transfer { get; set; }
         public bool? Ido { get; set; }
         public int? IdoSemester { get; set; }
